requirements.txt
ollama==0.3.3
streamlit==1.40.0
pyyaml==6.0.2
nltk==3.9.1